feat: implement space permission checks in item management and refactor related classes

+++ chiba-modules/chiba-storage/src/main/java/org/gragon/storage/service/impl/ItemServiceImpl.java
@@ -5,12 +5,14 @@ import com.baomidou.mybatisplus.core.toolkit.Wrappers;
 import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.gragon.common.core.exception.ServiceException;
 import org.gragon.common.core.utils.MapstructUtils;
 import org.gragon.common.core.utils.ObjectDiffUtils;
 import org.gragon.common.core.utils.StringUtils;
 import org.gragon.common.json.utils.JsonUtils;
 import org.gragon.common.mybatis.core.page.PageQuery;
 import org.gragon.common.mybatis.core.page.TableDataInfo;
+import org.gragon.common.satoken.utils.LoginHelper;
 import org.gragon.storage.domain.Item;
 import org.gragon.storage.domain.ItemOperationLog;
 import org.gragon.storage.domain.bo.ItemBo;
@@ -18,11 +20,12 @@ import org.gragon.storage.domain.vo.ItemVo;
 import org.gragon.storage.mapper.ItemMapper;
 import org.gragon.storage.mapper.ItemOperationLogMapper;
 import org.gragon.storage.service.ItemService;
-import org.gragon.storage.service.StorageSpaceService;
+import org.gragon.storage.service.SpacePermissionService;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.time.LocalDateTime;
+import java.util.List;
 import java.util.Map;
 
 @Service
@@ -32,7 +35,7 @@ public class ItemServiceImpl implements ItemService {
     private final ItemMapper baseMapper;
     private final ItemOperationLogMapper logMapper;
 
-    private final StorageSpaceService storageSpaceService;
+    private final SpacePermissionService spacePermissionService;
 
 
     /**
@@ -59,7 +62,14 @@ public class ItemServiceImpl implements ItemService {
 
     private Wrapper<Item> buildQueryWrapper(ItemBo item) {
         Map<String, Object> params = item.getParams();
+        // 权限校验
+        Long userId = LoginHelper.getUserId();
+        List<Long> accessibleSpaceIds = spacePermissionService.getAccessibleSpaceIds(userId);
+        // 如果用户没有任何空间权限，传递一个不存在的空间id
+        if (accessibleSpaceIds.isEmpty()) accessibleSpaceIds.add(-1L);
+
         return Wrappers.<Item>lambdaQuery()
+                .in(Item::getSpaceId, accessibleSpaceIds)
                 .like(StringUtils.isNotBlank(item.getName()), Item::getName, item.getName())
                 .between(params.get("beginTime") != null && params.get("endTime") != null,
                         Item::getCreateTime, params.get("beginTime"), params.get("endTime"))
@@ -73,9 +83,12 @@ public class ItemServiceImpl implements ItemService {
      * @return 插入结果
      */
     public int insertItem(ItemBo itemBo) {
+        if (!spacePermissionService.hasSpacePermission(LoginHelper.getUserId(), itemBo.getSpaceId())) {
+            throw new ServiceException("用户没有空间权限");
+        }
+
         Item item = MapstructUtils.convert(itemBo, Item.class);
-        item.setCreateTime(LocalDateTime.now());
-        item.setUpdateTime(LocalDateTime.now());
+        item.setOwnerId(LoginHelper.getUserId());
         // TODO 生成 (条形码、二维码) 编码
         return baseMapper.insert(item);
     }
@@ -88,15 +101,21 @@ public class ItemServiceImpl implements ItemService {
      */
     @Transactional(rollbackFor = Exception.class)
     public int updateItem(ItemBo itemBo) {
+        if (itemBo.getSpaceId() != null && !spacePermissionService.hasSpacePermission(LoginHelper.getUserId(), itemBo.getSpaceId())) {
+            throw new ServiceException("用户没有目标空间权限");
+        }
+
         Item item = baseMapper.selectById(itemBo.getId());
         if (item == null) return 0;
+
+        if (!spacePermissionService.hasSpacePermission(LoginHelper.getUserId(), itemBo.getSpaceId())) {
+            throw new ServiceException("用户没有空间权限");
+        }
 
         Item newItem = MapstructUtils.convert(itemBo, Item.class);
 
         ItemOperationLog log = new ItemOperationLog();
         ObjectDiffUtils.DiffNode<Item> node = ObjectDiffUtils.getDiffObjects(item, newItem, Item.class);
-
-        Long newStorageSpaceId = node.getNewObj().getSpaceId();
 
         log.setItemId(itemBo.getId());
         log.setNewItemData(JsonUtils.toJsonString(node.getNewObj(), true, true));
@@ -108,13 +127,17 @@ public class ItemServiceImpl implements ItemService {
     }
 
     /**
-     * 删除物品信息
+     * 删除物品信息 (进行逻辑删除)
      *
      * @param id 物品ID
      * @return 删除结果
      */
     public int deleteItem(Long id) {
-        // 逻辑删除
+        Item item = baseMapper.selectById(id);
+        if (item == null) return 0;
+        if (!spacePermissionService.hasSpacePermission(LoginHelper.getUserId(), item.getSpaceId())) {
+            throw new ServiceException("用户没有空间权限");
+        }
         return baseMapper.deleteById(id);
     }
 }
